pyproject.toml
[tool.poetry]
name = "qa-guru-20"
version = "0.1.0"
description = ""
authors = ["kris-zvereva <krs.zvrv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^9.0.2"
selene = ">=2.0.0rc4"
appium-python-client = "^5.2.4"
python-dotenv = "^1.2.1"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
allure-python-commons = "^2.15.2"
allure-pytest = "^2.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--clean-alluredir --alluredir=allure-results"
